Removed the counting code
I have removed the code that counts how many results there are. I have identified the issue with it, it needs access to the search:results data by being in the loop, but the issue is that because it is a loop the data replicates itself for the amount of entries it finds. There isn't access to a variable like count, so I have taken it out in the meantime