Changes to the industry icons
I have updated the taxonomy so that it now accepts an Image which is then displayed in the "orange block" on the products page. This way the user can choose what icon is shown by the taxonomy rather than having a fixed icon.

I have also updated the code to account for multisite urls